Send SitemapBuilder to its own class
